requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pydantic
pydantic-settings
httpx
pytest
requests
